% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assign.size.R
\name{assign.size}
\alias{assign.size}
\title{Assign spot sizes to 'rawDist' objects for estimating spatial extent of sample averaging error.}
\usage{
assign.size(rawDist, file = NULL, names = "generate.invalid",
  types = c("oval", "freehand", "rect"))
}
\arguments{
\item{rawDist}{\code{\link[=convert.ijdata]{rawDist}} object to which the values should be assigned.}

\item{file}{optional. ImageJ .zip file containing the spot size information. If \code{NULL}, the file is assumed to be the same than from where 'rawDist' \link[=read.ijdata]{data was read from}. The file must be located in the working directory.}

\item{names}{optional. A character argument specifying how the names of \code{spots} should be generated. See \code{\link{read.ijdata}} for details. Defaults to "generate.invalid".}

\item{types}{optional. A character vector specifying the \code{strType} of ROI objects to be considered as sample spots (see \code{\link{plot.ijroi}} for possible pattern types). Defaults to \code{c("oval", "freehand", "rect")} meaning that oval and freehand selections, as well as rectangle tool selections will be used to calculate the spatial extent of sample spots.}
}
\value{
Returns a list of class 'rawDist' with \code{\link{hyperframe}} containing spot size information.
}
\description{
Assigns spot sizes to \code{\link[=convert.ijdata]{rawDist}} objects for \link[=spot.dist]{estimating averaging error}.
}
\author{
Mikko Vihtakari
}
\seealso{
\code{\link{read.ijdata}} \code{\link{spot.dist}} \code{\link{assign.value}}
}

