% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/assign.value.R
\name{assign.value}
\alias{assign.value}
\title{Assign values to 'rawDist' objects for plotting.}
\usage{
assign.value(rawDist, value)
}
\arguments{
\item{rawDist}{\code{\link[=convert.ijdata]{rawDist}} object to which the values should be assigned.}

\item{value}{A list of \code{\link{data.frame}}s of the same length than spot spequences (\code{\link[=read.ijdata]{spots}}). Each \code{data.frame} has to have identical length to spots. First column specifies the order. Second column containing the values.}
}
\value{
Returns a list of class 'rawDist' containing spot value information.
}
\description{
Assigns values to \code{\link[=convert.ijdata]{rawDist}} objects for \link[=plot.rawDist]{spatial density plotting}.
}
\author{
Mikko Vihtakari
}

