% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{barium}
\alias{barium}
\title{Barium to calcium ratios for sample spots in shellspots dataset}
\format{A dataframe containing Ba/Ca information for \link{shellspots}. The object is ready for \code{\link{assign.size}} function.}
\usage{
data(barium)
}
\description{
Barium to calcium ratios for sample spots in shellspots dataset
}
\keyword{datasets}

