% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.ijdata.R
\name{convert.ijdata}
\alias{convert.ijdata}
\title{Convert IJDATA object to a list of spatstat.geom objects}
\usage{
convert.ijdata(X, Accuracy = 0.1)
}
\arguments{
\item{X}{an \code{\link[=read.ijdata]{IJDATA}} object to be converted.}

\item{Accuracy}{numeric value defining the number to round to. See \code{\link{round_any}}.}
}
\value{
Returns a list of class \code{rawDist}, which contains \link[spatstat.geom]{spatstat.geom} point patterns. The returned \code{rawDist} can be plotted using the generic plotting command.
}
\description{
Converts an \code{\link[=read.ijdata]{IJDATA}} to a list of \link[spatstat.geom]{spatstat.geom} patterns ready for \link[=plot.rawDist]{plotting} or \link[=spot.dist]{sample spot alignment}.
}
\examples{
data(shellspots)
shell_map <- convert.ijdata(shellspots)
plot(shell_map)
}
\seealso{
\code{\link{read.ijdata}} for generating \code{IJDATA} objects.

\code{\link{plot.rawDist}} for plotting.

\code{\link{spot.dist}} for aligning sample spots.
}
\author{
Mikko Vihtakari
}
