% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.dist.R
\name{growth.dist}
\alias{growth.dist}
\title{Calculate growth using a variety of measurement techniques}
\usage{
growth.dist(rawDist, coord.type = "scaled", type = "auto")
}
\arguments{
\item{rawDist}{a \code{\link[=convert.ijdata]{rawDist}} object for which the alignment should be done.}

\item{coord.type}{A character or numeric argument defining the coordinate type to do the calculations on. Alternatives:
\itemize{
  \item \strong{"scaled"} or \strong{1}. Only scaled coordinates.
  \item \strong{"original"} or any other character/number. Only original ImageJ ROI coordinates.
}}

\item{type}{}
}
\value{
A list of class \code{"growthDist"}. The first element (\code{$data}) contains a data.frame of growth measurements. The \code{$id} column gives the order of growth increments from the start point defined by the \code{main} axis. The \code{$gap} column gives the name of growth lines between which the growth is measured. Following columns give the growth measurements in units defined by the user in \code{\link{read.ijdata}} function (see the \code{scale} and \code{unit} arguments). The columns denoted as \code{.pr} give the percentage growth as compared to estimated total growth in the sample. The second list element contains \link[spatstat]{spatstat} segmented line (\code{psp}) patterns of the various growth measurement methods. List elements with \code{NULL} values were not applicable for the sample type. Abbreviations for measurement types are given in parenthesis. The types are:
\itemize{
  \item \code{$main} (main).
  \item \code{$manual} (man). Only applicable for \code{cross} section types.
  \item \code{$caliber} (cal). Only applicable for \code{along} section types.
  \item \code{$maximum} (max).
  \item \code{$maximum.along.x1} (maxx1). Only applicable for \code{cross} section types.
  \item \code{$maximum.along.x2} (maxx2). Only applicable for \code{cross} section types.
  \item \code{$guided.maximum} (maxg). Only applicable for \code{cross} section types.
  \item \code{$minimum} (min). Only applicable for \code{cross} section types.
  \item \code{$guided.minimum} (ming). Only applicable for \code{cross} section types.
  \item \code{$direct} (dire). Only applicable for \code{along} section types.
}
}
\description{
Calculates growth from rawDist \code{\link[=convert.ijdata]{rawDist}} objects and provides a range of results depending on the type of the sample.
}
\details{
This function is still under developemt and may not work as intended.
}
\seealso{
\code{\link{read.ijdata}} for reading zip files containing ImageJ ROIs.
\code{\link{convert.ijdata}} for converting the coordinate information to \link[spatstat]{spatstat} point patterns.
}
