% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ijdata.define.names.R
\name{ijdata.define.names}
\alias{ijdata.define.names}
\title{Define names for ROI data}
\usage{
ijdata.define.names(names, types, rois, data, type.names)
}
\arguments{
\item{names}{see the \code{\link[=read.ijdata]{names}} argument in \code{read.ijdata}}

\item{types}{a data.frame of ROI types. Calculated internally}

\item{rois}{integer vector defining the order of ROIs in \code{types} which will be affected by the function}

\item{data}{a data.frame containing the coordinate data. Calculated internally}

\item{type.names}{a character vector defining the manual names. See the \code{\link[=read.ijdata]{spot.names,gbs.names}} argument in \code{read.ijdata}.}
}
\value{
Returns a character vector of names
}
\description{
An internal function which defines names for various ROI objects
}
\author{
Mikko Vihtakari
}
\keyword{internal}
