% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ijdata.define.spots.R
\name{ijdata.define.spots}
\alias{ijdata.define.spots}
\title{Define the spot sequences from an ImageJ zip file}
\usage{
ijdata.define.spots(spots, types, scale, names, spot.names)
}
\arguments{
\item{spots}{a character specifying the type of ROI that should be considered as sampling spot sequences. See the argument with the same name in \code{\link{read.ijdata}}.}

\item{types}{a data.frame of ROI types. Calculated internally.}

\item{names}{a character argument. See the argument with the same name in \code{\link{read.ijdata}}.}

\item{spot.names}{a character vector specifying the manually added spot names. See the argument with the same name in \code{\link{read.ijdata}}.}

\item{rois}{a numeric value defining the scale of photograph in pixels / \code{unit}.}
}
\value{
Returns a list of x and y coordinates for samling sequences.
}
\description{
An internal function which defines spot sequences
}
\author{
Mikko Vihtakari
}
\keyword{internal}
