% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{ijzip}
\alias{ijzip}
\title{\code{ijzip} object from \code{\link[RImageJROI]{read.ijzip}} function (RImageJROI package)}
\format{\code{ijzip} object obtained using \code{\link[RImageJROI]{read.ijzip}} function from RImageJROI package. A list containing ImageJ ROI information about sample spots and growth lines along a bivalve mollusk shell section.}
\usage{
data(ijzip)
}
\description{
\code{ijzip} object from \code{\link[RImageJROI]{read.ijzip}} function (RImageJROI package)
}
\keyword{datasets}

