% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp.slope.intercept.R
\name{intercept.psp}
\alias{intercept.psp}
\title{Calculate intercept for a spatstat psp object}
\usage{
intercept.psp(x)
}
\arguments{
\item{x}{a \code{\link[spatstat]{psp}} object}
}
\value{
Returns a numeric vector containing y-intercepts of line segments in the same order than \code{x}
}
\description{
Calculates an intercept for each line segment in a psp object
}
\details{
The function assumes that each line segment has an infinite domain and follows the line equation $y = slope * x + intercept$.
}
\author{
Mikko Vihtakari
}
\keyword{internal}
