% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.rawDist.R
\name{plot.rawDist}
\alias{plot.rawDist}
\title{Plot rawDist object}
\usage{
\method{plot}{rawDist}(x, ..., sample.name = "keep", spot.type = "id",
  spot.size = 2, color.palette = colorRampPalette(c("blue", "cyan",
  "yellow", "red"), bias = 1)(100))
}
\arguments{
\item{x}{rawDist object}

\item{...}{Arguments to be passed to other methods, such as \link[=par]{graphical parameters}. Not implemented yet.}

\item{sample.name}{A character vector of length 1 specifying the sample name to be plotted as an overall title for the plot (\code{\link[=plot.default]{main}}). Defaults to \code{"keep"} meaning that the sample name will be extracted from the \code{\link[=convert.ijdata]{rawDist}} object.}

\item{spot.type}{levels "id", "value", "idvalue".}

\item{spot.size}{cex parameter at the moment}

\item{color.palette}{color palette for "value" and "idvalue". Passed to \code{\link{colorRampPalette}}.}
}
\description{
Plots a map of \code{\link[=convert.ijdata]{rawDist}} object.
}
\details{
Because the function plots a sample map, the aspect ratio is forced to 1 and cannot be changed. This might cause troubles when trying to set the axis limits. Try resizing the graphics window.
}
\examples{
data(shellspots)
shell_map <- convert.ijdata(shellspots)
plot(shell_map)
}
\author{
Mikko Vihtakari
}
\seealso{
\code{\link{convert.ijdata}} for converting the coordinate information to \link[spatstat]{spatstat} point patterns.

\code{\link{read.ijdata}} for reading zip files containing ImageJ ROIs.

\code{\link{plot.spotDist}} for plotting \code{\link[=spot.dist]{spotDist}} objects.

\code{\link[=plot]{plot.default}} and other methods; \code{\link[=points]{points}}, \code{\link[=lines]{lines}}, \code{\link[=par]{par}}.
}

