% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.spotDist.R
\name{plot.spotDist}
\alias{plot.spotDist}
\title{Plot spotDist object}
\usage{
\method{plot}{spotDist}(x, ..., sample.name = "keep",
  crossing.points = FALSE, size = FALSE)
}
\arguments{
\item{x}{\code{\link[=spot.dist]{spotDist}} object}

\item{...}{Arguments to be passed to other methods, such as \link[=par]{graphical parameters}. Not implemented yet.}

\item{sample.name}{A character vector of length 1 specifying the sample name to be plotted as an overall title for the plot (\code{\link[=plot.default]{main}}). Defaults to \code{"keep"} meaning that the sample name will be extracted from the \code{\link[=spot.dist]{spotDist}} object.}

\item{crossing.points}{Indicates whether the crossing points between sampling spot sequence traverses and growth lines should be plotted. Defaults to FALSE.}

\item{size}{optional. A logical vector of length 1 indicating whether actual spot size should be used in plotting. Defaults to FALSE.}
}
\description{
Plots a map of \code{\link[=spot.dist]{spotDist}} object.
}
\details{
Because the function plots a sample map, the aspect ratio is forced to 1 and cannot be changed. This might cause troubles when trying to set the axis limits. Try resizing the graphics window.
}
\examples{
data(shellspots)
shell_map <- convert.ijdata(shellspots)
x <- spot.dist(shell_map)
plot(x)
}
\author{
Mikko Vihtakari
}
\seealso{
\code{\link{spot.dist}} for aligning sample spots.

\code{\link{convert.ijdata}} for converting the coordinate information to \link[spatstat]{spatstat} point patterns.

\code{\link{read.ijdata}} for reading zip files containing ImageJ ROIs.

\code{\link{plot.rawDist}} for plotting \code{\link[=convert.ijdata]{rawDist}} objects.

\code{\link[=plot]{plot.default}} and other methods; \code{\link[=points]{points}}, \code{\link[=lines]{lines}}, \code{\link[=par]{par}}.
}

