% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp.slope.intercept.R
\name{pointsAlongLines}
\alias{pointsAlongLines}
\title{Find points that lie along line based on x- or y coordinates}
\usage{
pointsAlongLines(X = NULL, x = NULL, y = NULL)
}
\arguments{
\item{X}{A line segment pattern (object of class "psp").}

\item{x, y}{Known x or y coordinates given as a numeric vector or a ppp object. One of the parameters has to be \code{NULL}.}
}
\value{
A point pattern (object of class "ppp") in the same window as X.
}
\description{
Given a line segment pattern, find a series of points that lie along it based on x or y coordinates
}
\author{
Mikko Vihtakari
}
