% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ijdata.R
\name{read.ijdata}
\alias{read.ijdata}
\title{Read ImageJ zip file containing several ROI files and extract coordinate information.}
\usage{
read.ijdata(
  X,
  spots = NULL,
  gbs = NULL,
  main = NULL,
  growth = NULL,
  names = "generate.invalid",
  spot.names = NULL,
  gbs.names = NULL,
  main.name = "main",
  sample.name = NULL,
  scale = 1,
  unit = NULL
)
}
\arguments{
\item{X}{a character string defining the name (including extension) or file path of an ImageJ zip file. Alternatively an \code{\link[=read.ijzip]{ijzip}} object.}

\item{spots, gbs}{optional. Sample spots and growth lines, respectively. One of the following options:
\itemize{
  \item \strong{\code{NULL}} (default). Type of the ROI object will be chosen automatically. Leads to selection of "point" and "polyline" objects as sample spots and growth lines.
  \item \strong{numeric or character vector} specifying the order or the names of elements of ROI objects that should be assigned as the corresponding type. Might be useful in debugging.
  \item \strong{a character string} specifying the type of ROI objects that should be considered as the corresponding type. This option is included for package development and backwards compatibility.
}}

\item{main}{optional. Main (measurement) axis. Same options than for \code{spots,gbs}. Only one measurement axis per ImageJ .zip file is allowed. Defaults to \code{"line"}.}

\item{growth}{optional. A character string specifying the name of the "polyline" ROI that was used as a growth axis. Only if one was used, otherwise use \code{NULL}.}

\item{names}{optional. A character argument specifying how the names of \code{spots} and \code{gbs} should be generated. These names will be used in further functions (\code{\link{convert.ijdata}}, \code{\link{spot.dist}}). In general, it is adviced to use simple ROI names without special characters (for example \code{-} is not allowed in a ROI name; see 'Details'). Possible \code{names} options are:
\itemize{
\item \code{"generate.invalid"} (default). Uses the ROI object names, except when they are not valid \code{\link{data.frame}} column names. In the latter case sequential names will be generated.
\item \code{"generate"}. Generates sequential names for all elements.
\item \code{"keep"}. Uses the ROI object names, except when they are not valid \code{\link{data.frame}} column names. In the latter case \code{\link{make.names}} function will be used to generate \code{data.frame} combatible column names.
\item \code{"force.keep"}. Uses the ROI object names as they are in the .zip file. Using this option might cause problems in consequent functions and is not recommended.
\item \code{"manual"}. Names for \code{spots} and \code{gbs} are searched from \code{spot.names} and \code{gbs.names} arguments, respectively. If these arguments are not \code{NULL} and have the same length than \code{spots} or \code{gbs}, the given names are used for the elements. If the condition is not true, the names are generated following \code{"generate.invalid"}.
}}

\item{spot.names, gbs.names}{optional. A character vector of equal length to \code{spots}/\code{gbs} defining the names of sample spot sequences/growth bands.}

\item{main.name}{optional. A character string defining the name of the measurement axis (\code{main}). If \code{main.name = "keep"}, the ROI object name will be used (not recommended, see "Details"). Otherwise the name will be taken from the argument. Defaults to \code{"main"}.}

\item{sample.name}{optional. A character string defining the name of the sample. File name without the extension or alternatively \code{ijzip} object name is used when \code{NULL} (default).}

\item{scale}{optional. A numeric value defining the scale of photograph in pixels / \code{unit}. Defaults to 1.}

\item{unit}{optional. A charater string defining the unit of measurements. See \code{scale}.}
}
\value{
Returns an "IJDATA" object, which is a list of data frames containing the x and y coordinates for sampling spot sequences (\code{spots.x} and \code{spots.y}), growth bands (\code{gbs.x} and \code{gbs.y}), measurement axis (\code{main.x} and \code{main.y}) and growth axis (\code{growth.x} and \code{growth.y}) together with sample name, scaling factor and unit of measurement.
}
\description{
A wrapper function, which reads an ImageJ zip file containing a collection of ROI files and outputs a list of data frames ready for \code{\link{convert.ijdata}} function.
}
\details{
In order to minimize the amount of text to be typed by a user, ROI objects of type "point" (this includes the "Multi-point Tool" points) are considered as sample spot sequences (\code{spots}) by default. Further, all "polyline" objects are assumed as growth bands (\code{gbs}) and "line" objects as the measurement axis (\code{main}) resulting to that only one "line" object is allowed per .zip file using the default settings. Alternatively, the user can specify the \code{spots}, \code{gbs}, and \code{main} objects manually using the order of the ImageJ .zip file with the exception that \bold{only one measurement axis is allowed} per \code{\link[=convert.ijdata]{rawDist}} or \code{\link[=spot.dist]{spotDist}} object.

Punctuation characters other than \code{_} or \code{.} should not be used as names of \code{spots} or \code{gbs}, because they tend to confuse the internal \code{\link[base]{grep}} functions in \code{\link{spot.dist}} function. Hence it is adviced to use one of the options renaming invalid names of \code{spots} and \code{gbs} (\code{"generate.invalid"}, \code{"generate"}, \code{"keep"}).

Custom growth axis, if different from \code{main}, can be defined by using one named "polyline" in the ROI file and typing the name to \code{growth} argument. In such case, \code{main} is used to define the direction of growth, while the \code{growth} axis is used in growth measurements and sample spot alignment (distances are projected first to the \code{growth} axis and then transfered to the \code{main} axis).
}
\examples{
# Locate the example zip file
path <- file.path(system.file("extdata", package = "sclero"), "shellspots.zip")

# You can replace 'path' by 'Your_file_name.zip'
dat <- read.ijdata(path)
summary(dat)

## Works also for IJZIP objects
dat2 <- read.ijzip(path)
dat2 <- read.ijdata(dat2)
dat[!(dat \%in\% dat2)] # Only the sample name differs
}
\seealso{
\code{\link{order.ijdata}} for ordering and subsetting \code{read.ijdata} output.

\code{\link{convert.ijdata}} for converting the coordinate information to \link[spatstat]{spatstat} point patterns.

\code{\link{spot.dist}} for aligning sample spot sequences.

\code{\link[RImageJROI]{read.ijroi}} and \code{\link[RImageJROI]{read.ijzip}} for reading ImageJ ROI and .zip files.
}
\author{
Mikko Vihtakari
}
