% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sclero-package.R
\docType{package}
\name{sclero}
\alias{sclero}
\alias{sclero-package}
\title{Measure growth patterns and align sampling spots in photographs}
\description{
Provides functions to measure growth patterns and align sampling spots in chronologically deposited materials. The package is intended for the fields of geology, sclerochronology, and dentrochronology.
}
\author{
Mikko Vihtakari \email{mikko.vihtakari@gmail.com}
}

