% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp.slope.intercept.R
\name{slope.psp}
\alias{slope.psp}
\title{Calculate slopes for a spatstat psp object}
\usage{
slope.psp(x, orthogonal = FALSE)
}
\arguments{
\item{x}{a \code{\link[spatsta]{psp}} object}

\item{orthogonal}{logical indicating whether a slope for orthogonal line to \code{x} should be returned instead of slope for \code{x}}
}
\value{
Returns a numeric vector containing slopes of line segments in the same order than \code{x}
}
\description{
Calculates a slope for each line segment in a psp object
}
\details{
The function assumes that each line segment has an infinite domain and follows the line equation $y = slope * x + intercept$.
}
\author{
Mikko Vihtakari
}
\keyword{internal}
