% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spot.dist.R
\name{spot.dist}
\alias{spot.dist}
\title{Distance of sampling spots from margin along a measurement axis}
\usage{
spot.dist(rawDist, alignment = "closest", sample.name = NULL,
  run.ae = TRUE, use.centroids = TRUE)
}
\arguments{
\item{rawDist}{a \code{\link[=convert.ijdata]{rawDist}} object for which the alignment should be done.}

\item{alignment}{Should the sample spots be aligned according to the shortest distance to surrounding growth lines (default, \code{closest}) or according to the distance to the crossing points of the sample sequence traverse and growth lines (\code{traverse}, see 'Details' for further information.)}

\item{sample.name}{An optional parameter over-riding the sample name. If left empty the sample name will be passed from the previous steps (\code{\link{read.ijdata}}, \code{\link{convert.ijdata}})}

\item{run.ae}{a logical indicating whether to run averaging error estimation, if \link[=assign.size]{'rawDist contains spot size information}. Defaults to TRUE.}

\item{use.centroids}{a logical indicating whether to use centroids of spot.owins instead of spots, if \link[=assign.size]{'rawDist contains spot size information}. Defaults to TRUE.}
}
\value{
Returns a list of class \code{spotDist}, which contains \link[spatstat]{spatstat} point patterns. The returned \code{rawDist} can be plotted using the generic plotting command.
}
\description{
Scale the location of sampling spots to visible growth bands, project the sampling spots along a measurement axis (\code{main}) and calculate the distance from margin. Useful for locating high-resolution LA-ICP-MS or SIMS sample spots in samples with non-linear growth bands.
}
\details{
daa
}
\examples{
data(shellspots)
shell_map <- convert.ijdata(shellspots)
x <- spot.dist(shell_map)
plot(x)
}
\author{
Mikko Vihtakari
}
\seealso{
\code{\link{read.ijdata}} for reading zip files containing ImageJ ROIs.

\code{\link{order.ijdata}} for ordering and subsetting \code{read.ijdata} output.

\code{\link{convert.ijdata}} for converting the coordinate information to \link[spatstat]{spatstat} point patterns.

\code{\link{plot.spotDist}} for plotting.
}

