% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spotDist.R
\name{print.spotDist}
\alias{print.spotDist}
\title{Print \code{spotDist} objects}
\usage{
\method{print}{spotDist}(x, ...)
}
\arguments{
\item{x}{\code{spotDist} object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=spot.dist]{spotDist}} objects
}
\author{
Mikko Vihtakari
}
\seealso{
\code{\link{spot.dist}}
}

